requirements.txt
aiofiles
faiss_cpu
fastapi
langchain
langchain_community
langchain_huggingface
numpy
openai
pydantic_settings
PyPDF2
pytest
Requests
streamlit
pydantic
uvicorn
python-multipart
python-dotenv
sentence-transformers
transformers
accelerate
